Split errors out into a separate module
